Move defs only used in config to config
CMK-12298
